Merge branch 'Mercury-13.20.0' into Mercury-13.19.0-PLAT-7514

# Conflicts:
#	release-notes.md

+++ alpha/apps/kaltura/lib/myEntryUtils.class.php
@@ -701,8 +701,13 @@ class myEntryUtils
 		$contentPath = myContentStorage::getFSContentRootPath();
 			
 		$entry_status = $entry->getStatus();
+
+		$serveingVODfromLive = self::shouldServeVodFromLive($entry);
+		$entryLengthInMsec = $serveingVODfromLive ? $entry->getRecordedLengthInMsecs() : $entry->getLengthInMsecs();
 		 
 		$thumbName = $entry->getId()."_{$width}_{$height}_{$type}_{$crop_provider}_{$bgcolor}_{$quality}_{$src_x}_{$src_y}_{$src_w}_{$src_h}_{$vid_sec}_{$vid_slice}_{$vid_slices}_{$entry_status}";
+		if ($serveingVODfromLive && $vid_slices > 0)
+			$thumbName .= "_duration_" . $entryLengthInMsec;
 
 		if ($orig_image_path)
 			$thumbName.= '_oip_'.basename($orig_image_path);
@@ -775,7 +780,7 @@ class myEntryUtils
 			$vid_slice = 0;
 		
 		$cacheLockKey = "thumb-processing-resize".$finalThumbPath;
-		// creating the thumbnail is a very heavy operation prevent calling it in parallel for the same thubmnail for 5 minutes
+		// creating the thumbnail is a very heavy operation prevent calling it in parallel for the same thumbnail for 5 minutes
 		if ($cache && !$cache->add($cacheLockKey, true, 5 * 60))
 			KExternalErrors::dieError(KExternalErrors::PROCESSING_CAPTURE_THUMBNAIL);
 
@@ -790,6 +795,8 @@ class myEntryUtils
 		if ($entry->getType() == entryType::PLAYLIST)
 			myPlaylistUtils::updatePlaylistStatistics($entry->getPartnerId(), $entry);
 
+		if ($serveingVODfromLive)
+			$orig_image_path = null;
 
 		while($count--)
 		{
@@ -806,14 +813,13 @@ class myEntryUtils
 				(!file_exists($orig_image_path))
 				)
 			{
-
 				if ($vid_sec != -1) // a specific second was requested
 				{
-					$calc_vid_sec = min($vid_sec, floor($entry->getLengthInMsecs() / 1000));
+					$calc_vid_sec = min($vid_sec, floor($entryLengthInMsec / 1000));
 				}
 				else if ($vid_slices != -1) // need to create a thumbnail at a specific slice
 				{
-					$calc_vid_sec = floor($entry->getLengthInMsecs() / $vid_slices * min($vid_slice, $vid_slices) / 1000);
+					$calc_vid_sec = floor($entryLengthInMsec / $vid_slices * min($vid_slice, $vid_slices) / 1000);
 				}
 				else if ($entry->getStatus() != entryStatus::READY && $entry->getLengthInMsecs() == 0) // when entry is not ready and we don't know its duration
 				{
@@ -840,7 +846,7 @@ class myEntryUtils
 						KExternalErrors::dieError(KExternalErrors::PROCESSING_CAPTURE_THUMBNAIL);
 
 					$success = false;
-					if($multi && $packagerRetries)
+					if(($multi || $serveingVODfromLive) && $packagerRetries)
 					{
 						list($picWidth, $picHeight) = $shouldResizeByPackager ? array($width, $height) : array(null, null);
 						$destPath = $shouldResizeByPackager ? $capturedThumbPath . uniqid() : $capturedThumbPath;
@@ -912,6 +918,8 @@ class myEntryUtils
 
 					$convertedImagePath = myFileConverter::convertImage($orig_image_path, $processingThumbPath, $width, $height, $type, $bgcolor, true, $quality, $src_x, $src_y, $src_w, $src_h, $density, $stripProfiles, $thumbParams, $format,$forceRotation);
 				}
+				if ($thumbCaptureByPackager && file_exists($packagerResizeFullPath))
+					unlink($packagerResizeFullPath);
 			}
 
 
@@ -932,7 +940,7 @@ class myEntryUtils
 				++$vid_slice;
 			}
 
-			if ($thumbCaptureByPackager && $shouldResizeByPackager)
+			if ($thumbCaptureByPackager && $shouldResizeByPackager && $multi && file_exists($packagerResizeFullPath))
 				unlink($packagerResizeFullPath);
 
 			if ($isEncryptionNeeded)
@@ -990,6 +998,9 @@ class myEntryUtils
 
 	public static function captureThumbUsingPackager($entry, $capturedThumbPath, $calc_vid_sec, &$flavorAssetId, $width = null, $height = null)
 	{
+		if (myEntryUtils::shouldServeVodFromLive($entry))
+			return self::captureLiveThumbUsingPackager($entry, 'recording', $capturedThumbPath, $calc_vid_sec, $width, $height);
+		
 		$mappedThumbEntryTypes = array(entryType::PLAYLIST);
 		if(in_array($entry->getType(), $mappedThumbEntryTypes))
 			return self::captureMappedThumbUsingPackager($entry, $capturedThumbPath, $calc_vid_sec, $flavorAssetId, $width, $height);
@@ -997,6 +1008,34 @@ class myEntryUtils
 		return self::captureLocalThumbUsingPackager($entry, $capturedThumbPath, $calc_vid_sec, $flavorAssetId, $width, $height);
 	}
 
+	private static function captureLiveThumbUsingPackager(entry $entry, $liveType, $destThumbPath, $calc_vid_sec, $width = null, $height = null)
+	{
+		$packagerCaptureUrl = kConf::get('packager_local_live_thumb_capture_url', 'local', null);
+		if (!$packagerCaptureUrl)
+			return false;
+
+		$url = 'p/' . $entry->getPartnerId() . '/e/' . $entry->getId();
+		$dc = self::getLiveEntryDcId($entry->getRootEntryId(), EntryServerNodeType::LIVE_PRIMARY);
+		if (is_null($dc))
+			return false;
+		$packagerCaptureUrl = str_replace(array ( "{dc}", "{liveType}"), array ( $dc, $liveType) , $packagerCaptureUrl );
+
+		//TEMP CODE: currently live thumbnail with nginx support only offset from the end
+		$lenInSec = floor($entry->getRecordedLengthInMsecs() / 1000);
+		$offset = $calc_vid_sec - $lenInSec;
+		if ($offset >= 0)
+			return false;
+
+		return self::curlThumbUrlWithOffset($url, $offset, $packagerCaptureUrl, $destThumbPath, $width, $height);
+	}
+	
+	private static function getLiveEntryDcId($entryId, $type)
+	{
+		$entryServerNode = EntryServerNodePeer::retrieveByEntryIdAndServerType($entryId, $type);
+		if (!$entryServerNode)
+			return null;
+		return $entryServerNode->getDCId();
+	}
 
 	private static function captureMappedThumbUsingPackager($entry, $capturedThumbPath, $calc_vid_sec, &$flavorAssetId, $width, $height)
 	{
